requirements.txt
requests~=2.31.0
urllib3~=2.0.7
PyYAML~=6.0.1
pandas~=1.3.5
openpyxl~=3.1.3
tabulate~=0.9.0
matplotlib~=3.5.3
python-docx~=1.1.2
